backend/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
aiogram>=3.4.0
sqlalchemy[asyncio]>=2.0.25
asyncpg>=0.29.0
alembic>=1.13.0
boto3>=1.34.0
pydantic-settings>=2.1.0
python-jose[cryptography]>=3.3.0
python-multipart>=0.0.6
httpx>=0.26.0
openpyxl>=3.1.0
